requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
httpx==0.27.2
feedparser==6.0.11
qdrant-client==1.12.1
sentence-transformers==3.2.1
numpy==2.1.3
tenacity==9.0.0
streamlit==1.41.1
rich==13.9.4
python-dateutil==2.9.0.post0
